+++ lib/app.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'package:mawaqit/core/app_navigator.dart';
 import 'package:mawaqit/core/constants.dart';
 import 'package:mawaqit/core/theme/app_theme.dart';
 import 'package:mawaqit/data/models/app_settings.dart';
@@ -26,6 +27,7 @@ class MawaqitApp extends ConsumerWidget {
     return MaterialApp(
       title: AppConstants.appName,
       debugShowCheckedModeBanner: false,
+      navigatorKey: appNavigatorKey,
       theme: AppTheme.light(),
       darkTheme: AppTheme.dark(),
       themeMode: themeMode,